+++ include/avif/avif.h
@@ -1021,6 +1021,8 @@ struct avifCodecSpecificOptions;
 //   a combination of settings are tweaked to simulate this speed range.
 // * Some encoder settings can be changed after encoding starts. Changes will take effect in the next
 //   call to avifEncoderAddImage().
+// * If Exif metadata is provided and neither AVIF_TRANSFORM_IROT or AVIF_TRANSFORM_IMIR is set,
+//   transformFlags/irot/imir will be set accordingly to the Exif orientation, if parsed successfully.
 typedef struct avifEncoder
 {
     // Defaults to AVIF_CODEC_CHOICE_AUTO: Preference determined by order in availableCodecs table (avif.c)